[SYCL] Fix tests for multitile devices
