+++ SYCL/Basic/query_emulate_subdevice.cpp
@@ -1,0 +1,6 @@
+// RUN: %clangxx -fsycl -fsycl-targets=%sycl_triple %s -o %t.out
+// RUN: env CreateMultipleSubDevices=2 EnableTimestampPacket=1 \
+// RUN: NEOReadDebugKeys=1 SYCL_DEVICE_FILTER="gpu" %t.out
+
+// UNSUPPORTED: gpu-intel-dg1
+#include "query.hpp"
